refactor: class명 변경 및 전반적인 lotto game 진행 로직 구현
